pyproject.toml
[project]
name = "clawmark-mcp"
version = "0.1.0"
description = "MCP server for screenshot capture + intelligent annotation. Combines Playwright (web), mss (screen) and Tesseract OCR for pixel-perfect element detection."
authors = [{ name = "Hubert (rainkode)", email = "rainkode@hlsi.tech" }]
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }

dependencies = [
    "fastmcp>=2.0.0",
    "Pillow>=10.0.0",
    "mss>=9.0.1",                  # capture écran cross-platform (rapide)
    "playwright>=1.40.0",          # capture web + DOM
    "pytesseract>=0.3.10",         # OCR
    "pydantic>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "ruff>=0.1.0",
]

[project.scripts]
clawmark-mcp = "clawmark.server:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/clawmark"]

[tool.ruff]
line-length = 100
target-version = "py311"
